Added more template images, TemplateImage.h, TemplateImage.cpp
I got rid of the template images with a thin brush size and replaced it
with images with a thick brush size. I created TemplateImage.h and
TemplateImage.cpp. This is to store a template image and also store the
min and max number of strokes for each image. I also created a folder
for the template images.